parse user input
in chess.java i stopped setting figures manually, now the user can input
in the style of "Ka1" to set a king at a1.
First letter needs to be uppercase, second lower case, third character a
number. No error checking so far.

in figures.java there is no a function which gives the unicode chess
figures for the letters.

in board.java we parse the user input for now.