Fixing confusion between matrix and check board notation
Add a method to transform the coordinates from checkboard to matrix
notation. And _dont_ do that in "checkCoordinates". Which from now on
works with check board notation.
From the outside it should all be check board notation style.
setFigure() and getFigure are the only ones so far which translate the
coordinates into matrix style.

Should solve issue #10